face_recognition/requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy
Pillow
flask


# FastAPI and Web Framework
fastapi
uvicorn
python-multipart
websockets

# Security and Encryption
cryptography==41.0.7
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
pymongo==4.6.0
motor==3.3.2

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Monitoring and Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1